MT_Goo/requirements.txt
accounts==0.2.4
alembic==1.11.1
amqp==2.6.1
anyio==3.7.1
asgiref==3.7.2
attrs==23.1.0
bcrypt==4.0.1
billiard==3.6.4.0
cachetools==5.3.1
celery==4.4.7
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==7.1.2
cryptography==41.0.2
cssselect==1.2.0
cssutils==2.7.1
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
dnspython==2.4.1
ecdsa==0.18.0
email-validator==1.3.1
emails==0.5.15
exceptiongroup==1.1.2
fastapi==0.70.1
gunicorn==20.1.0
h11==0.9.0
httptools==0.1.2
idna==3.4
inflect==5.6.2
Jinja2==2.11.3
kombu==4.6.11
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
more-itertools==10.0.0
packaging==23.1
passlib==1.7.4
Pillow==10.0.0
pluggy==0.13.1
premailer==3.10.0
psycopg2-binary==2.9.6
py==1.11.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.12
pytest==5.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2023.3
raven==6.10.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
sqlparse==0.4.4
starlette==0.16.0
tenacity==6.3.1
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.11.8
uvloop==0.17.0
vine==1.3.0
wcwidth==0.2.6
websockets==8.1
